test(coding): split skill-load observation coverage
Separate executable lifecycle, identity, PATH, and cleanup cases from protocol, schema, replay, expiry, digest, and consumer cases. Share only the fixture support used by both modules and preserve all 17 original test methods plus the adjacent five-test suite.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: Preserve deterministic process cleanup and every skill-load assertion; change test organization only.
Rejected: SIZE_OK exception | the lifecycle and protocol responsibilities split cleanly below 250 pure LOC.
Confidence: high
Scope-risk: narrow
Directive: Keep skill-load lifecycle tests separate from protocol and replay tests as the surface grows.
Tested: 22 skill-load tests; 2 relevant CLI tests; test-name/source/assertion equality; Ruff; compileall; pure-LOC gate; git diff --check.
Not-tested: Full discovery and remote CI after this test-only split.
Plan: .omo/plans/omh-domain-skill-quality-upgrades.md
